Start using separate files for events
